requirements.txt
# python
python-dotenv

# data
ibis-framework[duckdb,polars,deltalake,bigquery] @ git+https://github.com/ibis-project/ibis.git@master

# viz
jupyter
#matplotlib
#seaborn
plotly
#altair @ git+https://github.com/altair-viz/altair.git@main
streamlit

# ml
scikit-learn

